pyproject.toml
[tool.poetry]
name = "duckdb-dbt"
version = "0.2.0"
description = "Production weather data pipeline with DuckDB, dbt, Apache Iceberg, and Polaris REST catalog"
authors = ["Chanukya Pekala <chanukya.pekala@gmail.com>"]
readme = "docs/README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
duckdb = ">=1.4.2"
dbt-duckdb = ">=1.10.0"
pandas = ">=2.0.0"
numpy = ">=1.24.0"
requests = ">=2.31.0"
pyyaml = ">=6.0"
python-dotenv = ">=1.0.0"
pyiceberg = {extras = ["pyarrow", "s3fs"], version = ">=0.7.0"}
pyarrow = ">=17.0.0"
sqlalchemy = ">=2.0.0"
flask = ">=3.1.2"
prefect = ">=2.14.0"
streamlit = ">=1.28.0"
plotly = ">=5.17.0"
altair = ">=5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-cov = ">=5.0.0"
ruff = ">=0.6.0"
mypy = ">=1.11.0"
pre-commit = ">=3.8.0"
ipython = ">=8.0.0"

[tool.poetry.scripts]
catalog-server = "src.catalog.rest_server:main"
fetch-stations = "src.ingestion.fetch_stations:main"
fetch-observations = "src.ingestion.fetch_observations:main"
load-observations = "src.ingestion.write_observations:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
